Fix cmp_ent_build() function; add example for a icu sw compression

+++ lib/common/cmp_entity.h
@@ -155,9 +155,9 @@ uint32_t cmp_ent_create(struct cmp_entity *ent, enum cmp_data_type data_type,
 			int raw_mode_flag, uint32_t cmp_size_byte);
 
 /* create a compression entity and set the header fields */
-size_t cmp_ent_build(struct cmp_entity *ent, uint32_t version_id,
-		     uint64_t start_time, uint64_t end_time, uint16_t model_id,
-		     uint8_t model_counter, struct cmp_cfg *cfg, int cmp_size_bits);
+uint32_t cmp_ent_build(struct cmp_entity *ent, uint32_t version_id,
+		       uint64_t start_time, uint64_t end_time, uint16_t model_id,
+		       uint8_t model_counter, struct cmp_cfg *cfg, int cmp_size_bits);
 
 /*
  * write the compression parameters from a compression configuration into the
